requirements.txt
Django==2.2.4
djangorestframework==3.9.4
Pillow==6.1.0
